Hack CMake to ignore the "build environment" prefix in its searches
In [conda-forge GitHub issue #24](https://github.com/conda-forge/casacore-feedstock/issues/24)
there was a problem because CMake decided to link with a version of the
ncurses library found in the "build environment" prefix, rather than the "host
environment" -- the major versions were different and so the resulting
binaries did not link against the major version intended in the recipe
specification.

As far as I can tell, the problem is that CMake decides that the "build
environment" library directory should be searched for shared libraries, which
is in turn because the build-env prefix shows up in the variable
`CMAKE_SYSTEM_PREFIX_PATH`. We hack the CMakeLists.txt file to remove the
relevant directory, which fixes the problem. As a bonus, a bunch of messages
about libraries like `libz` being shadoweded between the build and host
prefixes now go away.

The [CMake
docs](https://cmake.org/cmake/help/latest/variable/CMAKE_SYSTEM_PREFIX_PATH.html)
say that you're not supposed to futz with this variable, but I couldn't find
an alternative solution. It may be fixable by patching our cmake package --
although I don't have a good sense as to whether we might sometimes want the
build-env prefix to be in `CMAKE_SYSTEM_PREFIX_PATH`.