recipe/build.sh: update bison fix
We're still building with Bison 3.8 but now I need to set a new C/C++
preprocessor `#define` for each different parser variant name ... and
there are a lot of them. The long `$CPPFLAGS` variable is lame but I
don't know of an obvious way to get all the definitions into a header
that will (1) be included globally and (2) not end up polluting the
package artifacts.